third_party/WebKit/LayoutTests/crypto/subtle/aes-cbc/import-jwk-expected.txt

Test importing a JWK key for AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing a key...

PASS key.type is 'secret'
PASS key.extractable is false
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 256
PASS key.usages is ["encrypt"]

Using the key to encrypt plaintext...
PASS bytesToHexString(new Uint8Array(encryptionResult)) is 'f58c4c04d6e5f1ba779eabfb5f7bfbd69cfc4e967edb808d679f777bc6702c7d39f23369a9d9bacfa530e26304231461c9aaf02a6a54e9e242ccbf48c59daca6'
PASS successfullyParsed is true

TEST COMPLETE
